prototype_system/requirements.txt
beautifulsoup4==4.13.4
pylatexenc==2.10
PyYAML==6.0.2
regex==2024.11.6
Requests==2.32.4
tiktoken==0.9.0
toml==0.10.2
tqdm==4.67.1
aiohttp==3.11.16
async-lru==2.0.5
async-timeout==5.0.1
pandas==2.2.3
streamlit==1.46.1
streamlit-lottie
fpdf
streamlit_pdf_viewer
datetime